更新实体类

+++ src/main/java/com/example/schoolwall/entity/Answer.java
@@ -10,7 +10,7 @@ public class Answer {
     @TableId(value = "aid",type = IdType.AUTO)
     private int aid; //回答id
     private int qid;  //问题id
-    private int auid;  //回答者id
+    private long auid;  //回答者id
     private String acontent; //回答内容
     private Timestamp atime; //回答时间
     private Timestamp alikecount; //回答点赞次数
